second commit

+++ api/controllers/things/destroy-one-thing.js
@@ -19,6 +19,10 @@ module.exports = {
     forbidden: {
       description: 'The user does not own this thing',
       responseType: 'forbidden'
+    },
+    notFound: {
+      description: 'This thing does not exist',
+      responseType: 'notFound'
     }
   },
 
@@ -26,9 +30,14 @@ module.exports = {
   fn: async function (inputs) {
     var thing = await Thing.findOne({ id: inputs.id });
 
+    if (!thing) {
+      throw 'notFound';
+    }
+
     if (thing.user !== this.req.me.id) {
       throw 'forbidden';
     }
+
     await Thing.destroy({ id: inputs.id });
     return;
   }
